move menu components to new directory
